Create an octree for the purposes of collision. All quads up to the chunk depth go into the octree. Chunks also go into the octree but their collision is as of yet unimplemented. So far, traces against the octree are pretty fast. Constructing the octree needs some perf improvement.

+++ sp/src/sp_common.h
@@ -231,7 +231,7 @@ public:
 public:
 	bool				bHit;
 	bool				bStartInside;
-	float				flFraction;
+	double				flFraction;
 	CScalableVector		vecHit;
 	CScalableVector		vecNormal;
 };
